% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_fns.R
\name{.pkg_fns}
\alias{.pkg_fns}
\title{List functions exported by a package}
\usage{
.pkg_fns(pkg)
}
\arguments{
\item{pkg}{A character naming an installed package.}
}
\value{
A list of the actual functions exported by the package (not just
their names).
}
\description{
The trick is sorting out data from functions. There may be an easier way to
do this.
}
\keyword{internal}
