% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_fns.R
\name{pkg_fn_info}
\alias{pkg_fn_info}
\title{List exported functions and their arguments}
\usage{
pkg_fn_info(pkg)
}
\arguments{
\item{pkg}{A character naming an installed package.}
}
\value{
A named list of named pairlists. The names of the outer list are the
functions, and the names of each contained pairlist are the arguments. The
values of the pairlists are the default values of the arguments.
}
\description{
Generate a list of the functions exported by a package, along with their
arguments, and the default values of those arguments.
}
\details{
Note: While you can get values for "base" as if it is a package, primitive
functions will show \code{NULL} as their arguments.
}
\examples{
base_fns <- pkg_fn_info("base")
# Note that we do not know information about the arguments of primatives.
base_fns$is.function

pkg_fn_info("utils")
}
